+++ src/Accounts/Authentication/Factories/AuthenticationFactory.cs
@@ -488,6 +488,16 @@ namespace Microsoft.Azure.Commands.Common.Authentication.Factories
         /// <param name="environment"></param>
         public void RemoveUser(IAzureAccount account, IAzureEnvironment environment)
         {
+            RemoveUser(account, environment: null);
+        }
+
+        /// <summary>
+        /// Remove any stored credentials for the given user and the Azure environment used.
+        /// </summary>
+        /// <param name="account">The account to remove credentials for</param>
+        /// <param name="environment">The environment which account belongs to</param>
+        public void RemoveUser(IAzureAccount account, IAzureEnvironment environment)
+        {
             if (account != null && !string.IsNullOrEmpty(account.Id) && !string.IsNullOrWhiteSpace(account.Type))
             {
                 switch (account.Type)
@@ -513,10 +523,10 @@ namespace Microsoft.Azure.Commands.Common.Authentication.Factories
                             // make best effort to remove credentials
                         }
 
-                        RemoveFromTokenCache(account);
+                        RemoveFromTokenCache(account, environment.ActiveDirectoryAuthority);
                         break;
                     case AzureAccount.AccountType.User:
-                        RemoveFromTokenCache(account);
+                        RemoveFromTokenCache(account, environment.ActiveDirectoryAuthority);
                         break;
                 }
             }
@@ -558,7 +568,7 @@ namespace Microsoft.Azure.Commands.Common.Authentication.Factories
             return account.GetProperty(tokenKey);
         }
 
-        private void RemoveFromTokenCache(IAzureAccount account)
+        private void RemoveFromTokenCache(IAzureAccount account, string authority = null)
         {
             PowerShellTokenCacheProvider tokenCacheProvider;
             if (!AzureSession.Instance.TryGetComponent(PowerShellTokenCacheProvider.PowerShellTokenCacheProviderKey, out tokenCacheProvider))
@@ -566,12 +576,12 @@ namespace Microsoft.Azure.Commands.Common.Authentication.Factories
                 throw new NullReferenceException(Resources.AuthenticationClientFactoryNotRegistered);
             }
 
-            var publicClient = tokenCacheProvider.CreatePublicClient();
+            var publicClient = tokenCacheProvider.CreatePublicClient(authority);
             var accounts = publicClient.GetAccountsAsync()
                             .ConfigureAwait(false).GetAwaiter().GetResult();
             var tokenAccounts = accounts.Where(a => MatchCacheItem(account, a));
             foreach (var tokenAccount in tokenAccounts)
-                {
+            {
                 publicClient.RemoveAsync(tokenAccount)
                                 .ConfigureAwait(false).GetAwaiter().GetResult();
             }